cambios en widgets
widgets del sistema y personalizados se visualizan correctamente.

+++ rb-admin/core/files/file-del.php
@@ -8,15 +8,16 @@ require_once ABSPATH.'rb-script/funciones.php';
 
 $value_id = $_GET['id'];
 
-$q = $objDataBase->Ejecutar("SELECT src FROM photo WHERE id=$value_id");
+$q = $objDataBase->Ejecutar("SELECT src, type FROM photo WHERE id=$value_id");
 $r = $q->fetch_assoc();
 $src_img = $r['src'];
+$fileType = $r['type'];
 
 // eliminar foto fisica
 $ruta_image = ABSPATH.'rb-media/gallery/'.$src_img;
 $ruta_tn = ABSPATH.'rb-media/gallery/tn/'.$src_img;
 unlink($ruta_image);
-unlink($ruta_tn);
+if($fileType == "image/png" || $fileType == "image/jpeg" || $fileType == "image/gif") unlink($ruta_tn);
 
 $r = $objDataBase->Ejecutar("DELETE FROM photo WHERE id = $value_id");
